feat: complete overhaul pt 1
